// src/types/generated/config/ConfigDocument.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConfigDiagnostic } from "./ConfigDiagnostic";
import type { ConfigDocumentKind } from "./ConfigDocumentKind";
import type { ConfigScope } from "./ConfigScope";

export type ConfigDocument = { kind: ConfigDocumentKind, scope: ConfigScope, value: unknown, etag: string, readOnly: boolean, invalid: boolean, filePath: string, diagnostics: Array<ConfigDiagnostic>, };